refactor(shared-contracts): converge on #1437's club-id helper location (#1440)
#1446 (issue #1437) landed the promoted normalizer first, at
src/naming/clubId.ts beside the #1428 snapshot-file matcher. Two helpers
for one identity rule would be the exact defect this issue exists to fix,
so this takes that file verbatim as the base and extends it rather than
keeping a parallel src/identity/ copy:

  - normalizeClubId accepts `unknown` — the adopting sites read raw
    scraped records and parsed JSON (`row['Club Number']`,
    `parsed['clubId']`), so normalizing at the boundary replaces a
    String() cast at every call site
  - it strips CSV import debris ("'180", "Club 180") before comparing —
    tolerance FindAClubMerger has had since #429, moved here instead of
    surviving as a private third convention
  - findClubEntry: the same rule for object keys, for the club index
    (own keys only, keeping the #1112 prototype-member guard)

All-zeros semantics follow #1437's file, not the earlier draft: '0000'
is preserved, never collapsed. clubIdsMatch keeps both-directions
matching. The widened signature is source-compatible with #1437's
useClubHistory caller.
