+++ packages/shared-contracts/src/index.ts
@@ -168,17 +168,6 @@ export {
   type ValidationResult,
 } from './validation/validators.js'
 
-// Canonical club identity (#1440) — the ONE definition of "the same club",
-// shared by the transformer (write time), the diff engine, the collector's FAC
-// merge, the MCP reader and the frontend pages (read time). Three private
-// conventions used to coexist here; every disagreement degraded to an empty
-// state rather than an error.
-export {
-  normalizeClubId,
-  clubIdsMatch,
-  findClubEntry,
-} from './identity/clubId.js'
-
 // Canonical snapshot file naming (#1428) — the one matcher that decides what
 // `district_<id>.json` means, shared by scripts/ and collector-cli so the
 // daily-reports sidecar cannot be mistaken for a district snapshot.
@@ -192,3 +181,14 @@ export {
   type DistrictSnapshotObject,
   type DistrictSnapshotDateIndex,
 } from './naming/snapshotFileNames.js'
+
+// Canonical club-id comparison (#1437) — Toastmasters emits both `00009905`
+// and `9905`, so identity is the normalized form, never a strict `===`.
+// #1440 adopts this at the remaining call sites and adds `findClubEntry`, the
+// same rule for object keys (the club index is keyed by whichever form the
+// snapshot that generated it stored).
+export {
+  normalizeClubId,
+  clubIdsMatch,
+  findClubEntry,
+} from './naming/clubId.js'
